Add Column names to entity objects, changed fetch to eager, change popularity percentile to float

+++ src/main/java/com/nhlstenden/programming_language_api/models/LanguagePopularityDto.java
@@ -14,5 +14,5 @@ public class LanguagePopularityDto {
     private long id;
     private String languageName;
     private int year;
-    private int ratingPercentile;   //Percentage of projects using this language in that year
+    private float ratingPercentile;   //Percentage of projects using this language in that year
 }
